+++ android/App/src/test/kotlin/com/darakeon/dfm/summary/SummaryActivityTest.kt
@@ -54,8 +54,7 @@ class SummaryActivityTest: BaseTest() {
 
 	@Test
 	fun structure() {
-		activity.onCreate(null, null)
-		activity.waitTasks(mocker.server)
+		activity.onCreate(Bundle(), null)
 
 		assertNotNull(activity.findViewById(R.id.highlight))
 		assertNotNull(activity.findViewById(R.id.total_title))