+++ docs/RELEASES.md
@@ -21,7 +21,7 @@ Legend:
 
 ## <a name="4.5.0.1"></a>4.5.0.1 :ant: <sup>`6`</sup>
 - [x] `201124>201124` Fix contract pt-br word
-- [x] `201123>201123` Fix bug with android tests
+- [x] `201123>201127` Fix bug with android tests
 - [x] `201122>201122` Fix bug at future moves reports, showing zero value for move
 - [x] `201122>201123` Fix refresh gesture conflict with scroll of lists
 - [x] `201121>201122` Fix edit detailed move at site